fix: Accademic records for udpate and delete

+++ api/src/Modules/ProfessionalRecords/Applications/DeleteAcademicRecordCase.php
@@ -5,6 +5,7 @@ namespace Modules\ProfessionalRecords\Applications;
 use Illuminate\Support\Facades\DB;
 use Modules\ProfessionalRecords\Repositories\AcademicRecordRepository;
 use Modules\ProfessionalRecords\Services\ProfessionalFileStorageService;
+use Infrastructure\Exceptions\JsonResponseException;
 
 class DeleteAcademicRecordCase
 {
@@ -13,15 +14,17 @@ class DeleteAcademicRecordCase
         private ProfessionalFileStorageService $fileService
     ) {}
 
-    public function exec(int $recordId): void
+    public function exec(int $recordId, int $userId): void
     {
-        DB::transaction(function () use ($recordId) {
+        DB::transaction(function () use ($recordId, $userId) {
             $record = $this->repository->findByIdOrFail($recordId);
 
-            // Delete the file
+            if($record->user_id !== $userId){
+                throw new JsonResponseException('Unauthorized', 401);
+            }
+            
             $this->fileService->deleteFile($record->file);
 
-            // Delete the record
             $this->repository->delete($record);
         });
     }
